Added standard deviation, returning json with security data, removed …

+++ FinancialAPI/Stock.cs
@@ -3,12 +3,12 @@
     public class Stock
     {
         public DateTime Timestamp { get; set; }
-        public decimal Open { get; set; }
+        public double Open { get; set; }
 
-        public decimal High { get; set; }
-        public decimal Low { get; set; }
+        public double High { get; set; }
+        public double Low { get; set; }
 
-        public decimal Close { get; set; }
-        public decimal Volume { get; set; }
+        public double Close { get; set; }
+        public double Volume { get; set; }
     }
 }
